...En cour

+++ AuthJWT/Models/Panier.cs
@@ -11,10 +11,14 @@ namespace AuthJWT.Models
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public string Id { get; set; }
         public string ProductId { get; set; }
-        public int Quantity { get; set; }
+        public float Quantity { get; set; }
         public float TotalPrice { get; set; }
         public string UserId { get; set; }
         public Product Product { get; set; }
         public User User { get; set; }
+        //public float Price()
+        //{
+        //    return Quantity * Product.Prix;
+        //}
     }
 }
